feat(map): cut over /map to MapLibre, remove Leaflet route
Phase 7 destructive cutover. /map-next graduates to /map; the old
Leaflet-based +page.svelte and its private components (CommunityRail,
MapControls) are deleted. The four MapLibre control classes
(BasemapsControl, BoostToggleControl, DataRefreshControl,
NavButtonsControl) move with the +page.svelte, and the four cross-route
imports (MapSearchBar, MerchantDrawerHash, MerchantListPanel,
TileLoadingIndicator) become same-directory imports.

+layout.svelte's fullscreen + headerless allowlists drop /map-next.
Restored the original /map svelte:head (title "BTC Map", og:image,
twitter:title/image) and the screen-reader h1 that the prototype was
missing. Touched up a few /map-next references in lib comments;
viewport.ts and the BASEMAP_STORAGE_KEY ("btcmap-next-basemap")
still mention "-next" intentionally — the storage key is preserved
so existing users keep their basemap preference, and viewport.ts's
legacy Leaflet variant disappears entirely in the next commit when
the Leaflet helpers are deleted.

Verified at /map: 200 OK, MapLibre canvas, no leaflet container,
search bar present, og + twitter meta restored. /map-next now 404s.
