+++ src/routes/map/controls/BoostToggleControl.ts
