+++ src/routes/map/controls/NavButtonsControl.ts
